% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govinfo.R
\name{govinfo_collections}
\alias{govinfo_collections}
\title{List all GovInfo collections}
\usage{
govinfo_collections()
}
\value{
tibble: collectionCode, collectionName, packageCount, granuleCount
}
\description{
List all GovInfo collections
}
